Add a mock to test applyChildLayouts
